this doesn't complain anymore until compilation
